+++ src/main/java/roomescape/dto/request/ReservationAddRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 import java.time.LocalDate;
 import roomescape.domain.Reservation;
