+++ src/main/java/roomescape/dto/request/ThemeAddRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 import roomescape.domain.Theme;
