+++ src/main/java/roomescape/service/ReservationService.java
@@ -5,8 +5,8 @@ import org.springframework.stereotype.Service;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
-import roomescape.dto.ReservationAddRequest;
-import roomescape.dto.ReservationResponse;
+import roomescape.dto.request.ReservationAddRequest;
+import roomescape.dto.response.ReservationResponse;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
