+++ src/test/java/roomescape/service/FakeReservationTimeRepository.java
@@ -8,7 +8,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicLong;
 import roomescape.domain.ReservationTime;
-import roomescape.dto.ReservationTimeWithBookStatusResponse;
+import roomescape.dto.response.ReservationTimeWithBookStatusResponse;
 import roomescape.repository.ReservationTimeRepository;
 
 public class FakeReservationTimeRepository implements ReservationTimeRepository {
